# Better benchmarks using existing method