I'm not a huge fan of AMS, but it's a diservice to claim 25x when AMS, by default is underscored (because ruby / ActiveRecord is underscored, so telling to to perform additional transformations is cheating a bit).

Additionally, here are results with and without oj_mimic_json:

with
```
bundle exec rspec ./spec/lib/object_serializer_performance_spec.rb[1:3:4]
Run options: include {:ids=>{"./spec/lib/object_serializer_performance_spec.rb"=>["1:3:4"]}}

Serialize to JSON string 1000 with includes and meta
Serializer             Records    Time       Speed Up
Fast serializer        1000       28.13 ms  
AMS serializer         1000       439.81 ms  15.63x ✘
jsonapi-rb serializer  1000       98.9 ms    3.52x ✔
jsonapi-serializers    1000       133.2 ms   4.73x ✔

Serialize to Ruby Hash 1000 with includes and meta
Serializer             Records    Time       Speed Up
Fast serializer        1000       24.31 ms  
AMS serializer         1000       406.64 ms  16.73x ✘
jsonapi-rb serializer  1000       95.43 ms   3.93x ✔
jsonapi-serializers    1000       126.03 ms  5.18x ✔

```

without
```
bundle exec rspec ./spec/lib/object_serializer_performance_spec.rb[1:3:4]
Run options: include {:ids=>{"./spec/lib/object_serializer_performance_spec.rb"=>["1:3:4"]}}

Serialize to JSON string 1000 with includes and meta
Serializer             Records    Time       Speed Up
Fast serializer        1000       35.42 ms  
AMS serializer         1000       453.29 ms  12.8x ✘
jsonapi-rb serializer  1000       99.6 ms    2.81x ✔
jsonapi-serializers    1000       131.6 ms   3.72x ✔

Serialize to Ruby Hash 1000 with includes and meta
Serializer             Records    Time       Speed Up
Fast serializer        1000       23.02 ms  
AMS serializer         1000       428.47 ms  18.61x ✘
jsonapi-rb serializer  1000       94.08 ms   4.09x ✔
jsonapi-serializers    1000       127.92 ms  5.56x ✔

```